add qq version 8.4.1

+++ com.tencent.mobileqq/classes.jar/com/tencent/mobileqq/minigame/manager/GameGrowthGuardianManager$GuardInstructionDialogLogout$1.java
@@ -1,5 +1,6 @@
 package com.tencent.mobileqq.minigame.manager;
 
+import android.app.Activity;
 import android.content.DialogInterface;
 import android.content.DialogInterface.OnClickListener;
 import android.os.Process;
@@ -16,6 +17,11 @@ class GameGrowthGuardianManager$GuardInstructionDialogLogout$1
     paramDialogInterface.dismiss();
     try
     {
+      if ((this.this$0.getContext() != null) && ((this.this$0.getContext() instanceof Activity)))
+      {
+        ((Activity)this.this$0.getContext()).onBackPressed();
+        return;
+      }
       Process.killProcess(Process.myPid());
       return;
     }
@@ -27,7 +33,7 @@ class GameGrowthGuardianManager$GuardInstructionDialogLogout$1
 }
 
 
-/* Location:           L:\local\mybackup\temp\qq_apk\com.tencent.mobileqq\classes8.jar
+/* Location:           L:\local\mybackup\temp\qq_apk\com.tencent.mobileqq\classes9.jar
  * Qualified Name:     com.tencent.mobileqq.minigame.manager.GameGrowthGuardianManager.GuardInstructionDialogLogout.1
  * JD-Core Version:    0.7.0.1
  */
